Move ExampleCommand and ExampleSubsystem to templates folder
