+++ src/main/java/frc/robot/templates/ExampleCommand.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.utilities;
+package frc.robot.templates;
 
 import edu.wpi.first.wpilibj2.command.Command;
